Split ANSI terminal progress frame
Separate lifecycle state, row formatting, Unicode text layout, and VT100 rendering into focused files.


Copilot-Session: 6ccf776a-923f-486b-a99a-31a18f9433fb
